feat(#219): add Compiler and Decompiler abstractions
